requirements.txt
jupyter
matplotlib
opencv-python
pandas
pillow
pycocotools
nltk
numpy
requests
scikit-learn
scikit-image
scipy
seaborn
torch
torchvision
tqdm